Move css-template service in card service repository
