+++ ui/main/src/app/business/services/application-update.service.ts
@@ -11,7 +11,7 @@ import {Injectable} from '@angular/core';
 import {EntitiesService} from 'app/business/services/users/entities.service';
 import {GroupsService} from 'app/business/services/users/groups.service';
 import {LogOption, OpfabLoggerService} from 'app/business/services/logs/opfab-logger.service';
-import {TemplateCssService} from 'app/business/services/template-css.service';
+import {TemplateCssService} from 'app/business/services/card/template-css.service';
 import {UserService} from 'app/business/services/users/user.service';
 import {HandlebarsService} from 'app/business/services/card/handlebars.service';
 import {debounce, timer, map, catchError, switchMap} from 'rxjs';
